Graphviz moved to generic DiGraph
